requirements.txt
#Python 3.8.1
matplotlib==3.3.1
quandl-3.5.2
numpy==1.19.1
pandas==1.1.1